requirements.txt
fastapi
uvicorn
redis
requests
pyyaml
pyjwt
revChatGPT
loguru